Finished work on filtering processed data
Minor changes to requirements.txt and .gitignore
Changed custom warning messages in helpers.py. Work on it still in progress
There are big troubles with code performance. In future it's good idead to use numpy in to_numeric and to_datetime functions
To mitigate this problem serialization is added using pickle